Move more methods to IngestJobStatusFromJobTestData
