+++ java/clients/src/test/java/sleeper/clients/status/report/ingest/job/IngestJobStatusReporterTestData.java
@@ -22,7 +22,6 @@ import sleeper.core.tracker.ingest.job.IngestJobStatus;
 import sleeper.core.tracker.ingest.job.query.IngestJobAcceptedStatus;
 import sleeper.core.tracker.ingest.job.query.IngestJobRejectedStatus;
 import sleeper.ingest.core.job.IngestJob;
-import sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -36,14 +35,15 @@ import static sleeper.clients.status.report.StatusReporterTestHelper.job;
 import static sleeper.clients.status.report.StatusReporterTestHelper.task;
 import static sleeper.core.record.process.RecordsProcessedSummaryTestHelper.summary;
 import static sleeper.core.record.process.status.ProcessStatusUpdateTestHelper.defaultUpdateTime;
+import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.finishedIngestJob;
+import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.finishedIngestJobUncommitted;
 import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.ingestJobStatus;
 import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.ingestStartedStatus;
+import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.startedIngestJob;
 import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.acceptedRun;
 import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.acceptedRunWhichStarted;
 import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.failedIngestJob;
 import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.failedIngestRun;
-import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.finishedIngestJob;
-import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.finishedIngestJobUncommitted;
 import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.finishedIngestRun;
 import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.ingestAddedFilesStatus;
 import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.ingestFinishedStatusUncommitted;
@@ -128,7 +128,7 @@ public class IngestJobStatusReporterTestData {
                         .build()),
                 finishedIngestJobUncommitted(job5, task(3), summary(startTime5, Duration.ofMinutes(1), 600, 300), 3),
                 finishedIngestJob(job4, task(2), summary(startTime4, Duration.ofMinutes(1), 600, 300), 2),
-                IngestJobStatusFromJobTestData.startedIngestJob(job3, task(2), startTime3),
+                startedIngestJob(job3, task(2), startTime3),
                 finishedIngestJob(job2, task(1), summary(startTime2, Duration.ofMinutes(1), 600, 300), 1),
                 failedIngestJob(job1, task(1),
                         new ProcessRunTime(startTime1, Duration.ofMinutes(1)),
