+++ java/ingest/ingest-core/src/test/java/sleeper/ingest/core/job/status/IngestJobStatusFromJobTestData.java
@@ -15,6 +15,7 @@
  */
 package sleeper.ingest.core.job.status;
 
+import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.record.process.status.ProcessRun;
 import sleeper.core.tracker.ingest.job.IngestJobStatus;
 import sleeper.core.tracker.ingest.job.query.IngestJobAcceptedStatus;
@@ -54,6 +55,44 @@ public class IngestJobStatusFromJobTestData {
      */
     public static IngestJobStatus startedIngestJob(IngestJob job, String taskId, Instant startTime) {
         return IngestJobStatusTestHelper.ingestJobStatus(job.getId(), IngestJobStatusTestHelper.startedIngestRun(job, taskId, startTime));
+    }
+
+    /**
+     * Creates an ingest job status for a job that has finished.
+     *
+     * @param  job                  the ingest job
+     * @param  taskId               the ingest task ID
+     * @param  summary              the records processed summary
+     * @param  numFilesWrittenByJob the number of files written by the job
+     * @return                      an {@link IngestJobStatus}
+     */
+    public static IngestJobStatus finishedIngestJob(IngestJob job, String taskId, RecordsProcessedSummary summary, int numFilesWrittenByJob) {
+        return ingestJobStatus(job, IngestJobStatusTestHelper.finishedIngestRun(job, taskId, summary, numFilesWrittenByJob));
+    }
+
+    /**
+     * Creates an ingest job status for a job that has finished, but has not yet been committed to the state store.
+     *
+     * @param  job     the ingest job
+     * @param  taskId  the ingest task ID
+     * @param  summary the records processed summary
+     * @return         an {@link IngestJobStatus}
+     */
+    public static IngestJobStatus finishedIngestJobUncommitted(IngestJob job, String taskId, RecordsProcessedSummary summary) {
+        return finishedIngestJobUncommitted(job, taskId, summary, 1);
+    }
+
+    /**
+     * Creates an ingest job status for a job that has finished, but has not yet been committed to the state store.
+     *
+     * @param  job                  the ingest job
+     * @param  taskId               the ingest task ID
+     * @param  summary              the records processed summary
+     * @param  numFilesWrittenByJob the number of files written by the job
+     * @return                      an {@link IngestJobStatus}
+     */
+    public static IngestJobStatus finishedIngestJobUncommitted(IngestJob job, String taskId, RecordsProcessedSummary summary, int numFilesWrittenByJob) {
+        return ingestJobStatus(job, IngestJobStatusTestHelper.finishedIngestRunUncommitted(job, taskId, summary, numFilesWrittenByJob));
     }
 
     /**
